remove componentisation of bondingCurveGraph - not needed for v2-create
