+++ src/components/shared/formItems/BondingCurve/ProjectBondingCurveRate.tsx
@@ -13,7 +13,6 @@ import React, {
 
 import NumberSlider from '../../inputs/NumberSlider'
 import { FormItemExt } from '../formItemExt'
-import BondingCurveGraph from './BondingCurveGraph'
 
 export default function ProjectBondingCurveRate({
   name,
@@ -111,6 +110,16 @@ export default function ProjectBondingCurveRate({
     [calculator, graphCurve, value],
   )
 
+  const labelStyle: CSSProperties = {
+    fontSize: '.7rem',
+    fontWeight: 500,
+    textAlign: 'center',
+    position: 'absolute',
+  }
+
+  const graphSize = 180
+  const graphPad = 50
+
   return (
     <Form.Item
       name={name}
@@ -134,7 +143,61 @@ export default function ProjectBondingCurveRate({
       style={style}
       extra={
         <div style={{ display: 'flex', alignItems: 'center' }}>
-          <BondingCurveGraph />
+          <div style={{ position: 'relative' }}>
+            <div
+              style={{
+                display: 'flex',
+                alignItems: 'center',
+                justifyContent: 'center',
+                height: graphSize,
+                width: graphSize,
+              }}
+            >
+              <div
+                id={graphContainerId}
+                style={{
+                  width: graphSize - graphPad,
+                  height: graphSize - graphPad,
+                }}
+              ></div>
+            </div>
+
+            <div
+              style={{
+                position: 'absolute',
+                top: graphPad / 2,
+                left: graphPad / 2,
+                width: graphSize - graphPad,
+                height: graphSize - graphPad,
+                borderLeft: '2px solid ' + colors.stroke.secondary,
+                borderBottom: '2px solid ' + colors.stroke.secondary,
+              }}
+            ></div>
+
+            <div
+              style={{
+                ...labelStyle,
+                bottom: 0,
+                left: 0,
+                right: 0,
+              }}
+            >
+              % <Trans>tokens redeemed</Trans>
+            </div>
+
+            <div
+              style={{
+                ...labelStyle,
+                transform: 'rotate(-90deg)',
+                bottom: 0,
+                top: 0,
+                left: 0,
+                width: graphSize,
+              }}
+            >
+              <Trans>Token redeem value</Trans>
+            </div>
+          </div>
           <p>
             <Trans>
               This rate determines the amount of overflow that each token can be
